+++ SQLiteStudio3/guiSQLiteStudio/datagrid/sqlquerymodel.cpp
@@ -411,7 +411,7 @@ void SqlQueryModel::refreshGeneratedColumns(const QList<SqlQueryItem*>& items)
     for (auto resultIt = resultValues.begin(); resultIt != resultValues.end(); resultIt++)
     {
         SqlQueryItem* item = resultIt.key();
-        item->setValue(resultIt.value(), false, true);
+        item->setValue(resultIt.value(), true);
         item->setTextAlignment(findValueAlignment(resultIt.value(), item->getColumn()));
     }
 }
@@ -1088,16 +1088,8 @@ void SqlQueryModel::updateItem(SqlQueryItem* item, const QVariant& value, int co
 void SqlQueryModel::updateItem(SqlQueryItem* item, const QVariant& value, int columnIndex, const RowId& rowId, Qt::Alignment alignment)
 {
     SqlQueryModelColumnPtr column = columns[columnIndex];
-
-    // This should be equal at most, unless we have UTF-8 string, than there might be more bytes.
-    // If less, than it's not limited.
-    // An exception is when this is a column resulting from expression (function, operator, etc), because cell size limiting is applied
-    // only for columns resulting directly from table column. Cells without rowId and table are not limited to avoid slow "loadFullData()" calls
-    // on cells when - for example - copying full value of a cell.
-    bool limited = value.toByteArray().size() >= cellDataLengthLimit && !rowId.isEmpty() && column->editionForbiddenReason.isEmpty();
-
     item->setJustInsertedWithOutRowId(false);
-    item->setValue(value, limited, true);
+    item->setValue(value, true);
     item->setColumn(column.data());
     item->setTextAlignment(alignment);
     item->setRowId(rowId);
@@ -1174,7 +1166,7 @@ QHash<QString, QVariantList> SqlQueryModel::toValuesGroupedByColumns(const QList
 {
     QHash<QString, QVariantList> values;
     for (SqlQueryItem* item : items)
-        values[item->getColumn()->displayName] << item->getFullValue();
+        values[item->getColumn()->displayName] << item->getValue();
 
     return values;
 }
@@ -1191,7 +1183,6 @@ bool SqlQueryModel::readColumns()
     tablesInUse.clear();
 
     // Reading column mapping for ROWID columns
-    int totalRowIdCols = 0;
     AliasedTable aliasedTable;
     DbAndTable dbAndTable;
     for (const QueryExecutor::ResultRowIdColumnPtr& resCol : queryExecutor->getRowIdResultColumns())
@@ -1209,7 +1200,6 @@ bool SqlQueryModel::readColumns()
         aliasedTable.setTable(resCol->table);
         aliasedTable.setTableAlias(resCol->tableAlias);
         tableToRowIdColumn[aliasedTable] = resCol->queryExecutorAliasToColumn;
-        totalRowIdCols += resCol->queryExecutorAliasToColumn.size();
     }
 
     // Reading column details (datatype, constraints)
@@ -2092,56 +2082,6 @@ bool SqlQueryModel::isExecutionInProgress() const
     return queryExecutor->isExecutionInProgress();
 }
 
-void SqlQueryModel::loadFullDataForEntireRow(int row)
-{
-    int colCnt = columns.size();
-    SqlQueryItem *item = nullptr;
-    for (int col = 0; col < colCnt; col++)
-    {
-        item = itemFromIndex(row, col);
-        if (!item)
-            continue;
-
-        if (!item->isLimitedValue())
-            continue;
-
-        item->loadFullData();
-    }
-}
-
-void SqlQueryModel::loadFullDataForEntireColumn(int column)
-{
-    int rowCnt = rowCount();
-    SqlQueryItem *item = nullptr;
-    for (int row = 0; row < rowCnt; row++)
-    {
-        item = itemFromIndex(row, column);
-        if (!item)
-            continue;
-
-        if (!item->isLimitedValue())
-            continue;
-
-        item->loadFullData();
-    }
-}
-
-bool SqlQueryModel::doesColumnHaveLimitedValues(int column) const
-{
-    int rowCnt = rowCount();
-    SqlQueryItem *item = nullptr;
-    for (int row = 0; row < rowCnt; row++)
-    {
-        item = itemFromIndex(row, column);
-        if (!item)
-            continue;
-
-        if (item->isLimitedValue())
-            return true;
-    }
-    return false;
-}
-
 void SqlQueryModel::CommitUpdateQueryBuilder::clear()
 {
     database.clear();
@@ -2185,7 +2125,7 @@ QString SqlQueryModel::CommitUpdateQueryBuilder::build()
     int argIndex = 0;
     QString arg;
     QStringList assignments;
-    for (const QString& col : columns)
+    for (QString& col : columns)
     {
         arg = ":value_" + QString::number(argIndex++);
         assignmentArgs << arg;