+++ SQLiteStudio3/guiSQLiteStudio/datagrid/sqlqueryview.cpp
@@ -107,7 +107,6 @@ void SqlQueryView::createActions()
     createAction(INSERT_ROW, ICONS.INSERT_ROW, tr("Insert row"), this, SIGNAL(requestForRowInsert()), this);
     createAction(INSERT_MULTIPLE_ROWS, ICONS.INSERT_ROWS, tr("Insert multiple rows"), this, SIGNAL(requestForMultipleRowInsert()), this);
     createAction(DELETE_ROW, ICONS.DELETE_ROW, tr("Delete selected row"), this, SIGNAL(requestForRowDelete()), this);
-    createAction(LOAD_FULL_VALUES, ICONS.LOAD_FULL_VALUES, tr("Load full values"), this, SLOT(loadFullValuesForColumn()), this);
 
     actionMap[RESET_SORTING]->setEnabled(false);
 }
@@ -209,8 +208,6 @@ void SqlQueryView::setupHeaderMenu()
     headerContextMenu = new QMenu(horizontalHeader());
     headerContextMenu->addAction(actionMap[SORT_DIALOG]);
     headerContextMenu->addAction(actionMap[RESET_SORTING]);
-    headerContextMenu->addSeparator();
-    headerContextMenu->addAction(actionMap[LOAD_FULL_VALUES]);
 }
 
 QList<SqlQueryItem*> SqlQueryView::getSelectedItems()
@@ -319,11 +316,6 @@ void SqlQueryView::generateDelete()
     MAINWINDOW->openSqlEditor(getModel()->getDb(), sql);
 }
 
-void SqlQueryView::loadFullValuesForColumn()
-{
-    getModel()->loadFullDataForEntireColumn(headerContextMenuSection);
-}
-
 bool SqlQueryView::editInEditorIfNecessary(SqlQueryItem* item)
 {
     if (item->getColumn()->dataType.getType() == DataType::BLOB)
@@ -362,7 +354,7 @@ void SqlQueryView::paste(const QList<QList<QVariant> >& data)
             if (!validatePasting(warnedColumns, warnedRowDeletion, item))
                 continue;
 
-            item->setValue(theValue, false, false);
+            item->setValue(theValue, false);
         }
 
         return;
@@ -402,7 +394,7 @@ void SqlQueryView::paste(const QList<QList<QVariant> >& data)
             if (!validatePasting(warnedColumns, warnedRowDeletion, item))
                 continue;
 
-            item->setValue(cell, false, false);
+            item->setValue(cell, false);
         }
 
         // Go to next row, first column
@@ -527,7 +519,7 @@ void SqlQueryView::copy(bool withHeader)
     {
         for (SqlQueryItem* item : itemsInRows)
         {
-            itemValue = item->getFullValue();
+            itemValue = item->getValue();
             if (itemValue.userType() == QVariant::Double)
                 cells << doubleToString(itemValue);
             else
@@ -585,20 +577,6 @@ void SqlQueryView::scrollContentsBy(int dx, int dy)
     emit scrolledBy(dx, dy);
 }
 
-void SqlQueryView::mouseMoveEvent(QMouseEvent* event)
-{
-    QAbstractItemView::mouseMoveEvent(event);
-
-    QModelIndex idx = indexAt(QPoint(event->x(), event->y()));
-    if (idx != indexUnderCursor)
-    {
-        if (indexUnderCursor.isValid())
-            itemDelegate->mouseLeftIndex(indexUnderCursor);
-
-        indexUnderCursor = idx;
-    }
-}
-
 void SqlQueryView::updateCommitRollbackActions(bool enabled)
 {
     actionMap[COMMIT]->setEnabled(enabled);
@@ -628,11 +606,6 @@ void SqlQueryView::headerContextMenuRequested(const QPoint& pos)
 {
     if (simpleBrowserMode)
         return;
-
-    headerContextMenuSection = horizontalHeader()->visualIndexAt(pos.x());
-
-    bool hasLimitedValues = getModel()->doesColumnHaveLimitedValues(headerContextMenuSection);
-    actionMap[LOAD_FULL_VALUES]->setEnabled(hasLimitedValues);
 
     headerContextMenu->popup(horizontalHeader()->mapToGlobal(pos));
 }
@@ -794,7 +767,7 @@ void SqlQueryView::setNull()
         if (selItem->getColumn()->editionForbiddenReason.size() > 0)
             continue;
 
-        selItem->setValue(QVariant(QString()), false, false);
+        selItem->setValue(QVariant(QString()), false);
     }
 }
 
@@ -807,7 +780,7 @@ void SqlQueryView::erase()
         if (selItem->getColumn()->editionForbiddenReason.size() > 0)
             continue;
 
-        selItem->setValue("", false, false);
+        selItem->setValue("", false);
     }
 }
 
@@ -857,7 +830,7 @@ void SqlQueryView::openValueEditor(SqlQueryItem* item)
     MultiEditorDialog editor(this);
     editor.setWindowTitle(tr("Edit value"));
     editor.setDataType(item->getColumn()->dataType);
-    editor.setValue(item->getFullValue());
+    editor.setValue(item->getValue());
     editor.setReadOnly(!item->getColumn()->canEdit());
     if (editor.exec() == QDialog::Rejected)
         return;